metabench: Working for seq and unfactored.

+++ bin/metabench.py
@@ -3,13 +3,15 @@
 from __future__ import print_function
 import sys
 import os
+import subprocess
 import time
 import shutil
+import re
 import ConfigParser as cfgparser
-
-PYTHON = '/usr/bin/python2'
+from matopddl import PlanningProblem
 
 global_start_time = time.time()
+
 def p(*args):
     t = time.time()
     sys.stdout.write('[{0:.3f}] '.format(t - global_start_time))
@@ -114,13 +116,13 @@ class ConfigSearch(object):
     def __init__(self, cfg, name):
         self.name = name
         self.section = 'metabench-search-' + name
-        self.validate = cfg.getboolean(self.section, 'validate')
         self.type = cfg.get(self.section, 'type')
         self.search = cfg.get(self.section, 'search')
         self.heur = cfg.get(self.section, 'heur')
         self.max_time = cfg.getint(self.section, 'max-time')
         self.max_mem = cfg.getint(self.section, 'max-mem')
         self.repeat = cfg.getint(self.section, 'repeat')
+        self.cluster = cfg.get(self.section, 'cluster')
 
         self.bench = cfg.get(self.section, 'bench')
         self.bench = self.bench.split()
@@ -132,6 +134,7 @@ class ConfigSearch(object):
 
 class Config(object):
     def __init__(self, topdir):
+        topdir = os.path.abspath(topdir)
         self.topdir = topdir
         self.repodir = os.path.join(topdir, 'repo')
         self.search_bin = os.path.join(self.repodir, 'bin/search')
@@ -139,13 +142,13 @@ class Config(object):
         self.translate_py = os.path.join(self.repodir, 'third-party/translate/translate.py')
         self.protobuf_egg = os.path.join(self.repodir, 'third-party/protobuf/python/build/protobuf-2.6.1-py2.7.egg')
 
-        self.fn = os.path.join(topdir, 'config.cfg')
+        self.fn = os.path.join(topdir, 'metabench.cfg')
         self.cfg = cfgparser.ConfigParser()
         self.cfg.read(self.fn)
         self.repo = self.cfg.get('metabench', 'repo')
         self.commit = self.cfg.get('metabench', 'commit')
-        self.cplex_includedir = None
-        self.cplex_libdir = None
+        self.cplex_includedir = '/software/cplex-126/cplex/include'
+        self.cplex_libdir = '/software/cplex-126/cplex/lib/x86-64_linux/static_pic'
         if self.cfg.has_option('metabench', 'cplex-includedir'):
             self.cplex_includedir = self.cfg.get('metabench', 'cplex-includedir')
         if self.cfg.has_option('metabench', 'cplex-libdir'):
@@ -154,7 +157,6 @@ class Config(object):
         self.search = self.cfg.get('metabench', 'search')
         self.search = self.search.split()
         self.search = [ConfigSearch(self.cfg, x) for x in self.search]
-        print(self.cfg.items('metabench'))
 
 
 def repoClone(path, dst):
@@ -177,6 +179,7 @@ def repoMaplanMake(cfg):
     runCmd('cd {0} && make'.format(cfg.repodir))
     runCmd('cd {0} && make -C bin'.format(cfg.repodir))
 
+def repoCheck(cfg):
     check_files = [cfg.search_bin,
                    cfg.validate_bin,
                    cfg.translate_py,
@@ -194,6 +197,7 @@ def compileMaplan(cfg):
     repoClone(cfg.repo, cfg.repodir)
     repoCheckout(cfg.repodir, cfg.commit)
     repoMaplanMake(cfg)
+    repoCheck(cfg)
 
 
 def shEmpty(cfg, path):
@@ -207,13 +211,30 @@ def shTranslateSeq(cfg, path):
     sh = '''#!/bin/bash
 START=$(date +%s%N)
 cd {0};
+PYTHONPATH={2}:/software/python27-modules/software/python-2.7.6/gcc/lib/python2.7/site-packages \\
 {1} --proto --output {0}/problem.proto \\
     {0}/domain.pddl {0}/problem.pddl \\
         >{0}/translate.out 2>{0}/translate.err
 touch {0}/translate.done;
 END=$(date +%s%N)
 echo $(($END - $START)) >{0}/translate.nanotime
-'''.format(path, cfg.translate_py)
+'''.format(path, cfg.translate_py, cfg.protobuf_egg)
+    fout = open(os.path.join(path, 'translate.sh'), 'w')
+    fout.write(sh)
+    fout.close()
+
+def shTranslateUnfactor(cfg, path):
+    sh = '''#!/bin/bash
+START=$(date +%s%N)
+cd {0};
+PYTHONPATH={2}:/software/python27-modules/software/python-2.7.6/gcc/lib/python2.7/site-packages \\
+{1} --proto --output {0}/problem.proto \\
+    {0}/domain.pddl {0}/problem.pddl {0}/problem.addl \\
+        >{0}/translate.out 2>{0}/translate.err
+touch {0}/translate.done;
+END=$(date +%s%N)
+echo $(($END - $START)) >{0}/translate.nanotime
+'''.format(path, cfg.translate_py, cfg.protobuf_egg)
     fout = open(os.path.join(path, 'translate.sh'), 'w')
     fout.write(sh)
     fout.close()
@@ -222,8 +243,12 @@ def shValidateSeq(cfg, path):
     sh = '''#!/bin/bash
 START=$(date +%s%N)
 cd {0};
-{1} -v {0}/domain.pddl {0}/problem.pddl {0}/plan.out \\
-    >{0}/validate.out 2>{0}/validate.err
+if [ -f {0}/plan.out ]; then
+    {1} -v {0}/domain.pddl {0}/problem.pddl {0}/plan.out \\
+        >{0}/validate.out 2>{0}/validate.err
+else
+    echo "NO PLAN" >validate.out
+fi
 touch {0}/validate.done;
 END=$(date +%s%N)
 echo $(($END - $START)) >{0}/validate.nanotime
@@ -241,6 +266,28 @@ cd {0};
     -H {3} \\
     --max-time {4} \\
     --max-mem {5} \\
+    -o {0}/plan.out \\
+    --print-heur-init \\
+    >{0}/search.out 2>{0}/search.err
+touch {0}/search.done;
+END=$(date +%s%N)
+echo $(($END - $START)) >{0}/search.nanotime
+'''.format(path, cfg.search_bin, cfg_search.search, cfg_search.heur,
+           cfg_search.max_time, cfg_search.max_mem)
+    fout = open(os.path.join(path, 'search.sh'), 'w')
+    fout.write(sh)
+    fout.close()
+
+def shMaplanUnfactor(cfg, cfg_search, path):
+    sh = '''#!/bin/bash
+START=$(date +%s%N)
+cd {0};
+{1} -p {0}/problem.proto \\
+    -s {2} \\
+    -H {3} \\
+    --max-time {4} \\
+    --max-mem {5} \\
+    --ma-unfactor \\
     -o {0}/plan.out \\
     --print-heur-init \\
     >{0}/search.out 2>{0}/search.err
@@ -271,41 +318,71 @@ echo $(($END - $START)) >{0}/nanotime
 
 
 def createTasksSeq(topdir, cfg, cfg_search, cfg_bench, domain, problem, pddl):
+    if os.path.exists(os.path.join(topdir, 'domain.pddl')):
+        p('Skipping:', topdir, 'already created')
+        return;
+    p('Creating', topdir)
     shutil.copyfile(pddl['domain'], os.path.join(topdir, 'domain.pddl'))
     shutil.copyfile(pddl['problem'], os.path.join(topdir, 'problem.pddl'))
     shTranslateSeq(cfg, topdir)
-    if cfg_search.validate:
-        shValidateSeq(cfg, topdir)
-    else:
-        shEmpty(cfg, os.path.join(topdir, 'validate.sh'))
+    shValidateSeq(cfg, topdir)
     shMaplanSeq(cfg, cfg_search, topdir)
+    shSeq(cfg, cfg_search, topdir)
+
+def createTasksUnfactor(topdir, cfg, cfg_search, cfg_bench, domain, problem, pddl):
+    if os.path.exists(os.path.join(topdir, 'domain.pddl')):
+        p('Skipping:', topdir, 'already created')
+        return;
+    p('Creating', topdir)
+    pp = PlanningProblem(pddl['domain'], pddl['problem'])
+    pp.write_pddl_domain(os.path.join(topdir, 'domain.pddl'))
+    pp.write_pddl_problem(os.path.join(topdir, 'problem.pddl'))
+    pp.write_addl(os.path.join(topdir, 'problem.addl'))
+    pp.write_agent_list(os.path.join(topdir, 'problem.agents'))
+    shTranslateUnfactor(cfg, topdir)
+    shValidateSeq(cfg, topdir)
+    shMaplanUnfactor(cfg, cfg_search, topdir)
     shSeq(cfg, cfg_search, topdir)
 
 def createTasksProblem(topdir, cfg, cfg_search, cfg_bench, domain, problem):
     dst = os.path.join(topdir, problem)
-    os.mkdir(dst)
+    if not os.path.isdir(dst):
+        p('Creating', dst)
+        os.mkdir(dst)
     for repeat in range(cfg_search.repeat):
         dst2 = os.path.join(dst, str(repeat))
-        os.mkdir(dst2)
+        if not os.path.isdir(dst2):
+            os.mkdir(dst2)
         if cfg_search.type == 'seq':
             createTasksSeq(dst2, cfg, cfg_search, cfg_bench,
                            domain, problem, cfg_bench.pddl[domain][problem])
+        elif cfg_search.type in ['unfactored', 'unfactor']:
+            createTasksUnfactor(dst2, cfg, cfg_search, cfg_bench,
+                                domain, problem, cfg_bench.pddl[domain][problem])
+        else:
+            raise Exception('Unkown type {0}'.format(cfg_search.type))
 
 def createTasksDomain(topdir, cfg, cfg_search, cfg_bench, domain):
     dst = os.path.join(topdir, domain)
-    os.mkdir(dst)
+    if not os.path.isdir(dst):
+        p('Creating', dst)
+        os.mkdir(dst)
     for problem in cfg_bench.pddl[domain]:
         createTasksProblem(dst, cfg, cfg_search, cfg_bench, domain, problem)
 
 def createTasksSearchBench(topdir, cfg, cfg_search, cfg_bench):
     dst = os.path.join(topdir, cfg_bench.name)
-    os.mkdir(dst)
+    if not os.path.isdir(dst):
+        p('Creating', dst)
+        os.mkdir(dst)
     for domain in cfg_bench.pddl:
         createTasksDomain(dst, cfg, cfg_search, cfg_bench, domain)
 
 def createTasksSearch(cfg, cfg_search):
     dst = os.path.join(cfg.topdir, cfg_search.name)
-    os.mkdir(dst)
+    if not os.path.isdir(dst):
+        p('Creating', dst)
+        os.mkdir(dst)
     for cfg_bench in cfg_search.bench:
         createTasksSearchBench(dst, cfg, cfg_search, cfg_bench)
 
@@ -313,11 +390,82 @@ def createTasks(cfg):
     for cfg_search in cfg.search:
         createTasksSearch(cfg, cfg_search)
 
+
+
+
+pat_ma_agents = re.compile(r'^.*\(:agents +([a-zA-Z0-9-_ ]+)\).*$')
+def numAgents(addl):
+    with open(addl, 'r') as fin:
+        data = fin.read()
+        data = data.replace('\r\n', ' ')
+        data = data.replace('\n', ' ')
+        data = data.replace('\r', ' ')
+        match = pat_ma_agents.match(data)
+        if match is None:
+            p('Error: Could not parse `{0}\''.format(addl))
+        agents = match.group(1).split()
+        agents = filter(lambda x: len(x) > 0, agents)
+        return len(agents)
+
+def qsubCmd(cfg, cfg_search, root):
+    max_time = cfg_search.max_time + 600
+    max_mem = cfg_search.max_mem + 128
+    if cfg_search.cluster == 'luna':
+        resources = 'xeon:nodecpus16:cl_luna'
+        queue = '@arien'
+    elif cfg_search.cluster == 'zapat':
+        resources = 'nodecpus16:cl_zapat'
+        queue = '@wagap'
+
+    if cfg_search.type == 'seq':
+        ppn = 1
+        nodes = 1
+    elif cfg_search.type in ['unfactored', 'unfactor']:
+        ppn = numAgents(os.path.join(root, 'problem.addl'))
+        ppn = min(ppn + 1, 16)
+        nodes = 1
+    else:
+        raise Exception('TODO')
+
+    qargs = ['qsub']
+    qargs += ['-l', 'nodes={0}:ppn={1}:{2}'.format(nodes, ppn, resources)]
+    qargs += ['-l', 'walltime={0}s'.format(max_time)]
+    qargs += ['-l', 'mem={0}mb'.format(max_mem)]
+    qargs += ['-q', queue]
+    qargs += ['-o', os.path.join(root, 'task.out')]
+    qargs += ['-e', os.path.join(root, 'task.err')]
+    #qargs += ['-v', 'BENCH_DATA_FN={0},BENCH_TASK_I={1}'.format(bench_fn, task_i)]
+    qargs += [os.path.join(root, 'run.sh')]
+    return qargs
+
+def qsubTopdir(cfg, cfg_search, topdir):
+    for root, dirs, files in os.walk(topdir):
+        if 'run.sh' in files:
+            if 'qsub' in files:
+                p('qsub: Skipping', root)
+                continue
+
+            cmd = qsubCmd(cfg, cfg_search, root)
+            out = subprocess.check_output(cmd)
+            qsub_fn = os.path.join(root, 'qsub')
+            with open(qsub_fn, 'w') as fout:
+                p('{0} -> {1}'.format(out.strip(), qsub_fn))
+                fout.write(out)
+
+def qsub(cfg):
+    for cfg_search in cfg.search:
+        topdir = os.path.join(cfg.topdir, cfg_search.name)
+        qsubTopdir(cfg, cfg_search, topdir)
+
+
 def main(topdir):
     cfg = Config(topdir)
-    #compileMaplan(cfg)
+    if not os.path.exists(cfg.repodir):
+        compileMaplan(cfg)
+    else:
+        repoCheck(cfg)
     createTasks(cfg)
-    #translateSeq(cfg, '/home/danfis/tmp/metabench')
+    qsub(cfg)
     return 0
 
 if __name__ == '__main__':
